backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
Pillow==10.1.0
python-decouple==3.8
requests==2.31.0
anthropic==0.39.0
pdfplumber==0.11.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
